Merge

+++ mcrbin/stop.csh
@@ -1,6 +1,11 @@
 #!/bin/csh
 
-set id = `ps -ef | grep rpoldaemon | grep $USER | grep -v grep | awk '{ print $2}'`
-kill -s TERM $id
-echo "rpoldaemon stopped!"
-exit $?
+set id = `ps -e -u $USER | grep rpoldaemon | awk '{ print $1}'`
+if ( "$id" != "" ) then
+    kill -s TERM $id
+    echo "rpoldaemon stopped!"
+    exit 0
+else
+    echo "rpoldaemon is not running!"
+    exit 1
+endif
